Started to move GitHub-specific artifacts to github package
