+++ src/main/scala/com/atomist/source/git/github/domain/Commit.scala
@@ -1,0 +1,3 @@
+package com.atomist.source.git.github.domain
+
+case class Commit(sha: String)
