+++ src/main/scala/com/atomist/source/git/github/domain/CommitResponse.scala
@@ -1,0 +1,3 @@
+package com.atomist.source.git.github.domain
+
+case class CommitResponse(sha: String, url: String, tree: Ref, message: String, parents: Seq[Ref])
