+++ src/main/scala/com/atomist/source/git/github/domain/CreateTreeResponse.scala
@@ -1,0 +1,3 @@
+package com.atomist.source.git.github.domain
+
+case class CreateTreeResponse(sha: String, url: String, tree: Seq[TreeElement])
