+++ src/main/scala/com/atomist/source/git/github/domain/FileDeleteResponse.scala
@@ -1,0 +1,3 @@
+package com.atomist.source.git.github.domain
+
+case class FileDeleteResponse(commit: CommitResponse)
