+++ src/main/scala/com/atomist/source/git/github/domain/Links.scala
@@ -1,0 +1,5 @@
+package com.atomist.source.git.github.domain
+
+import com.fasterxml.jackson.annotation.JsonProperty
+
+case class Links(self: LinksHref, html: LinksHref, @JsonProperty("pull_request") pullRequest: LinksHref)
