+++ src/main/scala/com/atomist/source/git/github/domain/PullRequestMerged.scala
@@ -1,0 +1,3 @@
+package com.atomist.source.git.github.domain
+
+case class PullRequestMerged(sha: String, merged: Boolean, message: String)
