+++ src/main/scala/com/atomist/source/git/github/domain/PullRequestRequest.scala
@@ -1,0 +1,3 @@
+package com.atomist.source.git.github.domain
+
+case class PullRequestRequest(title: String, head: String, base: String, body: String)
