+++ src/main/scala/com/atomist/source/git/github/domain/Ref.scala
@@ -1,0 +1,3 @@
+package com.atomist.source.git.github.domain
+
+case class Ref(url: String, sha: String)
