+++ src/main/scala/com/atomist/source/git/github/domain/TreeElement.scala
@@ -1,0 +1,3 @@
+package com.atomist.source.git.github.domain
+
+case class TreeElement(path: String, mode: String, `type`: String, size: Int, sha: String, url: String)
