+++ src/test/scala/com/atomist/source/git/GitHubServicesTest.scala
@@ -3,8 +3,8 @@ package com.atomist.source.git
 import com.atomist.source._
 import com.atomist.source.file.NamedFileSystemArtifactSourceIdentifier
 import com.atomist.source.git.GitHubArtifactSourceLocator.MasterBranch
-import com.atomist.source.git.GitHubServices.PullRequestRequest
 import com.atomist.source.git.TestConstants._
+import com.atomist.source.git.github.domain.PullRequestRequest
 import com.atomist.util.GitRepositoryCloner
 import org.kohsuke.github.{GHIssueState, GHRepository}
