+++ src/test/scala/com/atomist/source/git/GitServicesTest.scala
@@ -1,10 +1,9 @@
 package com.atomist.source.git
 
 import com.atomist.source._
-import com.atomist.source.file.NamedFileSystemArtifactSourceIdentifier
 import com.atomist.source.git.GitHubArtifactSourceLocator.MasterBranch
-import com.atomist.source.git.GitHubServices.PullRequestRequest
 import com.atomist.source.git.TestConstants.Token
+import com.atomist.source.git.github.domain.PullRequestRequest
 import com.atomist.util.GitRepositoryCloner
 import org.kohsuke.github.GHRepository
